ログイン画面と新規登録画面の追加に伴う変更
