+++ lib/l10n/app_en.arb
@@ -21,8 +21,18 @@
   "calorie_count": "Calorie Count",
   "faq": "FAQ",
   "app_version": "App Version",
+
   "swap": "Switch accounts",
   "logout": "Logout",
+  "logout_confirmation": "Logout confirmation",
+  "are_you_sure_logout": "Are you sure you want to log out?",
+  "cancel": "Cancel",
+
+  "login": "Log in",
+  "password": "Password",
+  "password_confirmation": "Password confirmation",
+
+  "sign_up": "Sign up",
 
   "walker": "Walker",
   "runner": "Runner",
