+++ lib/l10n/app_ko.arb
@@ -21,8 +21,18 @@
   "calorie_count": "칼로리 계산",
   "faq": "자주 묻는 질문",
   "app_version": "앱 버전",
+
   "swap": "계정 전환",
   "logout": "로그아웃",
+  "logout_confirmation": "로그아웃 확인",
+  "are_you_sure_logout": "로그아웃하시겠습니까?",
+  "cancel": "취소",
+
+  "login": "로그인",
+  "password": "비밀번호",
+  "password_confirmation": "비밀번호 확인",
+
+  "sign_up": "신규 등록",
 
   "walker": "보행자",
   "runner": "러너",
