+++ lib/src/screens/home.dart
@@ -8,6 +8,7 @@ class HomeScreen extends StatelessWidget {
     return Scaffold(
       appBar: AppBar(
         title: const Text('アルケール'),
+        automaticallyImplyLeading: false,
       ),
       body: Stack(
         children: [
