+++ lib/src/screens/third.dart
@@ -8,6 +8,7 @@ class ThirdScreen extends StatelessWidget {
     return Scaffold(
       appBar: AppBar(
         title: const Text('3'),
+        automaticallyImplyLeading: false,
       ),
       body:
       const Center(child: Text('3rd画面', style: TextStyle(fontSize: 32.0))),
